+++ test/fixtures/mixed_version.rb
@@ -1,0 +1,5 @@
+module Reissue
+  module MixedGem
+    VERSION = "2.35.number19"
+  end
+end