+++ test/fixtures/prerelease_version.rb
@@ -1,0 +1,5 @@
+module Reissue
+  module PrereleaseGem
+    VERSION = "3.2.1.rc1"
+  end
+end