+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Reissue::Task class to handle the creation of reissue tasks
 - Return version and date from Reissue.finalize
 - Ability to limit the number of versions to maintain
+- Gem release support
 
 ### Fixed: