+++ CHANGELOG.md
@@ -5,12 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.1.6] - Unreleased
+## [0.1.6] - 2024-06-10
 
 ### Added
 
-- The ability to specify the version_redo_proc to customize the version number
-  update process.
+- The ability to specify the version_redo_proc to customize the version number update process.
 
 ### Fixed